split out validator code; update accordingly
